+++ libs/ui/src/lib/services/data.service.ts
@@ -589,6 +589,10 @@ export class DataService {
             item.date = parseISO(item.date);
           }
 
+          for (const item of data.splits ?? []) {
+            item.date = parseISO(item.date);
+          }
+
           return data;
         })
       );